refactor: rename main.py to impl.py and introduce dynamic main.py entry point [skip-gen]
